[#5] reset table

+++ database/migrations/2023_05_23_022906_create_farms_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('farms', function (Blueprint $table) {
             $table->id();
             $table->string('codename');
-            $table->integer('latitude');
-            $table->integer('longitude');
+            $table->float('latitude');
+            $table->float('longitude');
             $table->string('plant');
             $table->unsignedBigInteger('map_id');
             $table->foreign('map_id')->references('id')->on('maps')->onDelete('cascade');
